moved token exchange tests
