+++ api_tests/tests/oauth/test_oauth_endpoints.py
@@ -1,12 +1,7 @@
-from api_tests.scripts import config
+from api_tests.scripts.config import OAUTH_URL
 from api_tests.scripts.response_bank import BANK
 import pytest
 import random
-import requests
-import jwt
-from uuid import uuid4
-from time import time, sleep
-import json
 
 
 @pytest.mark.asyncio
@@ -146,7 +141,7 @@ class TestOauthEndpoints:
         # Make second callback request with same state value
         assert helper.check_endpoint(
             verb="GET",
-            endpoint=f"{config.OAUTH_BASE_URI}/{config.OAUTH_PROXY}/callback",
+            endpoint=f"{OAUTH_URL}/callback",
             expected_status_code=400,
             expected_response={
                 "error": "invalid_request",
@@ -745,460 +740,3 @@ class TestOauthEndpoints:
             endpoint="userinfo",
             headers={'Authorization': f'Bearer {self.oauth.access_token}'}
         )
-
-    @pytest.mark.happy_path
-    @pytest.mark.token_exchange
-    async def test_token_exchange_happy_path(self):
-        # Given
-        expected_status_code = 200
-        expected_expires_in = '599'
-        expected_token_type = 'Bearer'
-        expected_issued_token_type = 'urn:ietf:params:oauth:token-type:access_token'
-
-        id_token_jwt = self.oauth.create_id_token_jwt()
-        client_assertion_jwt = self.oauth.create_jwt(kid='test-1')
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            _jwt=client_assertion_jwt,
-            id_token_jwt=id_token_jwt
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code'], resp['body']
-        assert 'access_token' in resp['body']
-        assert expected_expires_in == resp['body']['expires_in']
-        assert expected_token_type == resp['body']['token_type']
-        assert expected_issued_token_type == resp['body']['issued_token_type']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_invalid_client_assertion_type(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing or invalid client_assertion_type - " \
-                                     "must be 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer"
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'Invalid',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token'
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_invalid_subject_token_type(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "missing or invalid subject_token_type - " \
-                                     "must be 'urn:ietf:params:oauth:token-type:id_token'"
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token_type': 'Invalid',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange'
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_claims_assertion_invalid_kid(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing 'kid' header in JWT"
-
-        client_assertion_jwt = self.oauth.create_jwt(kid=None)
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_claims_assertion_invalid_typ_header(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Invalid 'typ' header in JWT - must be 'JWT'"
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1", headers={'typ': 'invalid'})
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_claims_assertion_invalid_iss_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing or non-matching iss/sub claims in JWT"
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1", claims={
-            "sub": '',
-            "jti": str(uuid4()),
-            "aud": f"{config.OAUTH_URL}/token",
-            "exp": int(time()) + 5,
-        })
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_claims_assertion_missing_jti_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing jti claim in JWT"
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1", claims={
-            "sub": self.oauth.client_id,
-            "iss": self.oauth.client_id,
-            "jti": '',
-            "aud": f"{config.OAUTH_URL}/token",
-            "exp": int(time()) + 5,
-        })
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_claims_assertion_missing_exp_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing exp claim in JWT"
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1", claims={
-            "sub": self.oauth.client_id,
-            "iss": self.oauth.client_id,
-            "jti": str(uuid4()),
-            "aud": f"{config.OAUTH_URL}/token",
-        })
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_claims_assertion_invalid_exp_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Invalid exp claim in JWT - more than 5 minutes in future"
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1", claims={
-            "sub": self.oauth.client_id,
-            "iss": self.oauth.client_id,
-            "jti": str(uuid4()),
-            "aud": f"{config.OAUTH_URL}/token",
-            "exp": int(time()) + 50000,
-        })
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_claims_assertion_invalid_jti_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Non-unique jti claim in JWT"
-
-        id_token_claims = {
-            'at_hash': 'tf_-lqpq36lwO7WmSBIJ6Q',
-            'sub': '787807429511',
-            'auditTrackingId': '91f694e6-3749-42fd-90b0-c3134b0d98f6-1546391',
-            'amr': ['N3_SMARTCARD'],
-            'iss': 'https://am.nhsint.ptl.nhsd-esa.net:443/'
-                   'openam/oauth2/realms/root/realms/NHSIdentity/realms/Healthcare',
-            'tokenName': 'id_token',
-            'aud': '969567331415.apps.national',
-            'c_hash': 'bc7zzGkClC3MEiFQ3YhPKg',
-            'acr': 'AAL3_ANY',
-            'org.forgerock.openidconnect.ops': '-I45NjmMDdMa-aNF2sr9hC7qEGQ',
-            's_hash': 'LPJNul-wow4m6Dsqxbning',
-            'azp': '969567331415.apps.national',
-            'auth_time': 1610559802,
-            'realm': '/NHSIdentity/Healthcare',
-            'exp': int(time()) + 600,
-            'tokenType': 'JWTToken',
-            'iat': int(time()) - 10
-        }
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1")
-        id_token_jwt = self.oauth.create_id_token_jwt(kid="identity-service-tests-1", claims=id_token_claims)
-
-        # When
-        await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token': id_token_jwt,
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Second request should fail
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token': id_token_jwt,
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_subject_token_missing_iss_or_sub_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing or non-matching iss/sub claims in JWT"
-
-        id_token_claims = {
-            'at_hash': 'tf_-lqpq36lwO7WmSBIJ6Q',
-            'sub': '787807429511',
-            'auditTrackingId': '91f694e6-3749-42fd-90b0-c3134b0d98f6-1546391',
-            'amr': ['N3_SMARTCARD'],
-            'tokenName': 'id_token',
-            'aud': '969567331415.apps.national',
-            'c_hash': 'bc7zzGkClC3MEiFQ3YhPKg',
-            'acr': 'AAL3_ANY',
-            'org.forgerock.openidconnect.ops': '-I45NjmMDdMa-aNF2sr9hC7qEGQ',
-            's_hash': 'LPJNul-wow4m6Dsqxbning',
-            'azp': '969567331415.apps.national',
-            'auth_time': 1610559802,
-            'realm': '/NHSIdentity/Healthcare',
-            'exp': int(time()) + 600,
-            'tokenType': 'JWTToken',
-            'iat': int(time()) - 10
-        }
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1")
-        id_token_jwt = self.oauth.create_id_token_jwt(kid="identity-service-tests-1", claims=id_token_claims)
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token': id_token_jwt,
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_subject_token_missing_aud_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing aud claim in JWT"
-
-        id_token_claims = {
-            'at_hash': 'tf_-lqpq36lwO7WmSBIJ6Q',
-            'sub': '787807429511',
-            'auditTrackingId': '91f694e6-3749-42fd-90b0-c3134b0d98f6-1546391',
-            'iss': 'https://am.nhsint.ptl.nhsd-esa.net:443'
-                   '/openam/oauth2/realms/root/realms/NHSIdentity/realms/Healthcare',
-            'amr': ['N3_SMARTCARD'],
-            'tokenName': 'id_token',
-            'c_hash': 'bc7zzGkClC3MEiFQ3YhPKg',
-            'acr': 'AAL3_ANY',
-            'org.forgerock.openidconnect.ops': '-I45NjmMDdMa-aNF2sr9hC7qEGQ',
-            's_hash': 'LPJNul-wow4m6Dsqxbning',
-            'azp': '969567331415.apps.national',
-            'auth_time': 1610559802,
-            'realm': '/NHSIdentity/Healthcare',
-            'exp': int(time()) + 600,
-            'tokenType': 'JWTToken',
-            'iat': int(time()) - 10
-        }
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1")
-        id_token_jwt = self.oauth.create_id_token_jwt(kid="identity-service-tests-1", claims=id_token_claims)
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token': id_token_jwt,
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
-
-    @pytest.mark.errors
-    @pytest.mark.token_exchange
-    async def test_token_exchange_subject_token_missing_exp_claim(self):
-        # Given
-        expected_status_code = 400
-        expected_error = 'invalid_request'
-        expected_error_description = "Missing exp claim in JWT"
-
-        id_token_claims = {
-            'at_hash': 'tf_-lqpq36lwO7WmSBIJ6Q',
-            'sub': '787807429511',
-            'auditTrackingId': '91f694e6-3749-42fd-90b0-c3134b0d98f6-1546391',
-            'amr': ['N3_SMARTCARD'],
-            'iss': 'https://am.nhsint.ptl.nhsd-esa.net:443'
-                   '/openam/oauth2/realms/root/realms/NHSIdentity/realms/Healthcare',
-            'tokenName': 'id_token',
-            'aud': '969567331415.apps.national',
-            'c_hash': 'bc7zzGkClC3MEiFQ3YhPKg',
-            'acr': 'AAL3_ANY',
-            'org.forgerock.openidconnect.ops': '-I45NjmMDdMa-aNF2sr9hC7qEGQ',
-            's_hash': 'LPJNul-wow4m6Dsqxbning',
-            'azp': '969567331415.apps.national',
-            'auth_time': 1610559802,
-            'realm': '/NHSIdentity/Healthcare',
-            #'exp': int(time()) + 600,
-            'tokenType': 'JWTToken',
-            'iat': int(time()) - 10
-        }
-
-        client_assertion_jwt = self.oauth.create_jwt(kid="test-1")
-        id_token_jwt = self.oauth.create_id_token_jwt(kid="identity-service-tests-1", claims=id_token_claims)
-
-        # When
-        resp = await self.oauth.get_token_response(
-            grant_type="token_exchange",
-            data={
-                'grant_type': 'urn:ietf:params:oauth:grant-type:token-exchange',
-                'subject_token_type': 'urn:ietf:params:oauth:token-type:id_token',
-                'client_assertion_type': 'urn:ietf:params:oauth:client-assertion-type:jwt-bearer',
-                'subject_token': id_token_jwt,
-                'client_assertion': client_assertion_jwt
-            }
-        )
-
-        # Then
-        assert expected_status_code == resp['status_code']
-        assert expected_error == resp['body']['error']
-        assert expected_error_description == resp['body']['error_description']
